滑轨调试完成，底盘能动

+++ Core/Module/chassis.h
@@ -1,6 +1,6 @@
 #ifndef CHASSIS_H
 #define CHASSIS_H
-#include "can_motor.h"
+#include "motor.h"
 #include "pid.h"
 
 extern struct CAN_Motor can1_motor_1;
@@ -10,8 +10,8 @@ extern struct CAN_Motor can1_motor_4;
 
 struct Chassis_t
 {
-    int fbSpeed;
-    int lrSpeed;
+    float fbSpeed;
+    float lrSpeed;
     int rotate_speed;
 
     struct PID_t chassis_x_pid;
@@ -24,10 +24,11 @@ struct Chassis_t
 };
 
 
-void chassis_move(int x, int y);  //  
+void chassis_move(float x, float y);  //  
 void chassis_rotate(int angle);   //相对角度
 void chassis_navigate(int x,int y); //到达时定位（初步考虑低速运行）
 void chassis_brake(void);
+
 
 extern struct Chassis_t chassis;
